+++ src/main/java/usecases/actionsExecutor/ActionsExecutor.java
@@ -1,4 +1,4 @@
-package usecases.actions;
+package usecases.actionsExecutor;
 
 import entities.action.Action;
 import entities.action.ActionPaste;
