+++ src/main/java/usecases/shortcutsEvaluator/ShortcutsEvaluator.java
@@ -1,4 +1,4 @@
-package usecases.shortcuts;
+package usecases.shortcutsEvaluator;
 
 import java.util.ArrayList;
